requirements.txt
streamlit
langchain
openai
chromadb
python-dotenv
pypdf
pandas
openpyxl
phoenix